backend/requirements/base.txt
Django==2.0.4
djangorestframework==3.8.2
django-filter==2.1.0
django-environ==0.4.5
django-extensions==2.1.6
django-allauth==0.39.1
psycopg2
psycopg2-binary==2.7.7
djangorestframework-jwt
django-cors-headers
